% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carpark_datasets}
\alias{carpark_datasets}
\alias{Avon_Street_CP}
\alias{Charlotte_Street_CP}
\alias{Lansdown_P+R}
\alias{Newbridge_P+R}
\alias{Odd_Down_P+R}
\alias{Podium_CP}
\alias{SouthGate_General_CP}
\alias{SouthGate_Rail_CP}
\title{Individual car park datasets}
\format{Data frames with between 99225 and 201880 records of 7 variables.}
\usage{
Avon_Street_CP

Charlotte_Street_CP

Lansdown_P+R

Newbridge_P+R

Odd_Down_P+R

Podium_CP

SouthGate_General_CP

SouthGate_Rail_CP
}
\description{
All records from individual car parks from October 2014 to December 2016.
}
\keyword{datasets}
