% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{events}
\alias{events}
\title{Number of advertised events in Bath by date}
\format{Data frame with 823 records of 2 variables:
\describe{
 \item{Date}{date-time object}
 \item{count}{number of listed events}
}}
\usage{
events
}
\description{
The number of events listed on \url{http://www.bath.co.uk/events/} for each
 day in October 2014 through to December 2016.
}
\keyword{datasets}
