% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_records.R
\name{get_range_crude}
\alias{get_range_crude}
\title{Download records from a specified date range from the Bath: Hacked datastore}
\usage{
get_range_crude(from = NA, to = NA)
}
\arguments{
\item{from}{Datetime object for the earliest record to retrieve.}

\item{to}{Datetime object for the latest record to retrieve.}
}
\value{
Car parking records from the specified date range.
}
\description{
Retrieve raw records uploaded to the datastore within a specified date range.
}
\examples{
library(lubridate)

# Records for June 2016
raw_data <- get_range_crude(ymd_hms("2016-06-01 00:00:00"),
                            ymd_hms("2016-06-30 23:59:59"))

# All records since 14:30 on 1st January 2017
raw_data <- get_range_crude(from = ymd_hms("2017-01-01 14:30:00"))

# All records from before 2015
raw_data <- get_range_crude(to = ymd_hms("2014-12-31 23:59:59"))


# Retrieve all records (identical to get_all_crude)
raw_data <- get_range_crude()
}
\seealso{
\code{\link{get_all_crude}}
}

