% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{month_datasets}
\alias{month_datasets}
\alias{2014_10}
\alias{2014_11}
\alias{2014_12}
\alias{2015_01}
\alias{2015_02}
\alias{2015_03}
\alias{2015_04}
\alias{2015_05}
\alias{2015_06}
\alias{2015_07}
\alias{2015_08}
\alias{2015_09}
\alias{2015_10}
\alias{2015_11}
\alias{2015_12}
\alias{2016_01}
\alias{2016_02}
\alias{2016_03}
\alias{2016_04}
\alias{2016_05}
\alias{2016_06}
\alias{2016_07}
\alias{2016_08}
\alias{2016_09}
\alias{2016_10}
\alias{2016_11}
\alias{2016_12}
\title{Individual month datasets}
\format{Data frames with approximately 50000 records of 7 variables.}
\usage{
2014_10

2014_11

2014_12

2015_01

2015_02

2015_03

2015_04

2015_05

2015_06

2015_07

2015_08

2015_09

2015_10

2015_11

2015_12

2016_01

2016_02

2016_03

2016_04

2016_05

2016_06

2016_07

2016_08

2016_09

2016_10

2016_11

2016_12
}
\description{
Records for all car parks from individual months from October 2014 to
 December 2016. Each month contains approximately 50000 records (except
 \code{2014_10}).
}
\keyword{datasets}
