% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_records.R
\name{refine.deduplicate}
\alias{refine.deduplicate}
\title{Remove duplicate records (internal method)}
\usage{
refine.deduplicate(x, first_upload)
}
\arguments{
\item{x}{Data frame containing records.}

\item{first_upload}{If TRUE, ensures record with oldest DateUploaded is kept.}
}
\description{
(Internal method)
}
