% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather}
\alias{weather}
\title{Weather for 2014-10 to 2016-12}
\format{Data frame with 803 records of 8 variables:
\describe{
 \item{Date}{date of record}
 \item{MeanTemperatureC}{mean air temperature in degrees Celsius}
 \item{MeanHumidity}{mean air humidity, as a percentage}
 \item{MeanWindSpeedKmh}{mean wind speed in km/h}
 \item{Precipitationmm}{daily rainfall in mm}
 \item{Rain}{logical indicating presence of rain}
 \item{Fog}{logical indicating presence of fog}
 \item{Snow}{logical indicating presence of snow}
}}
\usage{
weather
}
\description{
Daily weather records for Bath from
 \href{https://www.wunderground.com/history/airport/EGTG/}{wunderground.com}
 for 2014-10-17 to 2016-12-27
}
\keyword{datasets}

